backends/monolith/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
cryptography==41.0.3
dateparser==1.1.1
distro==1.4.0
Django==4.1.10
django-cors-headers==4.2.0
django-multitenant==3.2.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
entrypoints==0.3
frozenlist==1.3.1
gevent==22.10.1
greenlet==1.1.3
gunicorn==20.1.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
multidict==6.0.2
netifaces==0.10.4
numpy==1.23.3
oauthlib==3.1.0
pandas==1.5.0
pbr==5.10.0
pexpect==4.6.0
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==23.2.0
pyrsistent==0.15.5
pyserial==3.4
python-binance==1.0.16
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
regex==2022.3.2
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
ssh-import-id==5.10
testresources==2.0.1
Twisted==21.2.0
tzdata==2022.4
tzlocal==4.2
ujson==5.5.0
urllib3==1.25.8
wadllib==1.3.3
websockets==10.3
yarl==1.8.1
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
